feature: se realiza la refactorizacion completa del Mid de acuerdo a las propuestas. udistrital/Cumplidos_y_certificaciones_documentacion#118
