--- helpers/helpers_estado_pago/refactor_helper.go
@@ -1,14 +1,0 @@
-package helpers_estado_pago
-
-import (
-	"encoding/json"
-)
-
-func LimpiezaRespuestaRefactor(respuesta map[string]interface{}, v interface{}) {
-	b, err := json.Marshal(respuesta["Data"])
-	if err != nil {
-		panic(err)
-	}
-
-	json.Unmarshal(b, v)
-}
