--- helpers/helpers_soporte/refactor_helper.go
@@ -1,14 +1,0 @@
-package helpers_soporte
-
-import (
-	"encoding/json"
-)
-
-func LimpiezaRespuestaRefactor(respuesta map[string]interface{}, v interface{}) {
-	b, err := json.Marshal(respuesta["Data"])
-	if err != nil {
-		panic(err)
-	}
-
-	json.Unmarshal(b, v)
-}
