+++ models/cumplido_proveedor.go
@@ -12,18 +12,16 @@ type CumplidoProveedor struct {
 }
 
 type SolicituRevisionCumplidoProveedor struct {
-	Id                 int
-	TipoContrato       string
-	NumeroContrato     string
-	VigenciaContrato   int
-	Rp                 string
-	NombreProveedor    string
-	Dependencia        string
-	Cdp                string
-	NombreOrdenador    string
-	DocumentoOrdenador string
-	VigenciaRP         string
-	FechaModificacion  time.Time
-	FechaCreacion      time.Time
-	Activo             bool
+	Id                int
+	TipoContrato      string
+	NumeroContrato    string
+	VigenciaContrato  int
+	Rp                string
+	NombreProveedor   string
+	Dependencia       string
+	Cdp               string
+	VigenciaRP        string
+	FechaModificacion time.Time
+	FechaCreacion     time.Time
+	Activo            bool
 }
